@@ -2,7 +2,7 @@ import {RegistrationService} from './../../services/registration.service';
 import {Component, EventEmitter, OnInit, Output} from '@angular/core';
 import {Router} from '@angular/router';
 import {HttpErrorResponse} from '@angular/common/http';
-import {FormBuilder, FormGroup, Validators} from '@angular/forms';
+import {UntypedFormBuilder, UntypedFormGroup, Validators} from '@angular/forms';
 
 @Component({
   selector: 'app-registration',
@@ -12,11 +12,11 @@ import {FormBuilder, FormGroup, Validators} from '@angular/forms';
 export class RegistrationComponent implements OnInit {
   registrationError: string;
 
-  public registrationForm: FormGroup;
+  public registrationForm: UntypedFormGroup;
 
   constructor(
     private registrationService: RegistrationService,
-    private router: Router, private formBuilder: FormBuilder) {
+    private router: Router, private formBuilder: UntypedFormBuilder) {
     this.registrationForm = this.formBuilder.group({
       email: this.formBuilder.control('', [Validators.required, Validators.email]),
       firstName: this.formBuilder.control('', [Validators.required]),
